wasmparser: detect "malformed" cases in parser alone (without validator)
This moves the bulk of the expression-reading logic into OperatorsReader
(out of BinaryReader).
