+++ tests/cli/dangling_if.wat.stderr
@@ -1,0 +1,1 @@
+error: control frames remain at end of function body or expression (at offset 0x1a)
