+++ tests/cli/dangling_if.wat.stdout
@@ -2,6 +2,3 @@
   (type (;0;) (func))
   (func (;0;) (type 0)
     if ;; label = @1
-    
-  )
-)
