+++ tests/cli/missing-features/missing-exceptions.wast
@@ -1,13 +1,13 @@
 ;; RUN: wast --assert default --snapshot tests/snapshots % -f wasm1
 
-(assert_malformed
+(assert_invalid
   (module binary
     "\00asm" "\01\00\00\00"
     "\0d\01\00"
     )
   "exceptions proposal not enabled")
 
-(assert_malformed
+(assert_invalid
   (module
     (import "" "" (tag))
     )
