+++ tests/cli/print-dont-reserve-the-world.wat.stderr
@@ -1,1 +1,1 @@
-error: module contains 4294967295 functions which exceeds the limit of 1000000
+error: function section has non-zero count but code section is absent (at offset 0xf)
