+++ tests/cli/print-no-panic-dangling-else.wat
@@ -1,4 +1,4 @@
-;; RUN: print %
+;; FAIL: print %
 
 (module
     (func else)
