+++ tests/cli/print-no-panic-dangling-else.wat.stderr
@@ -1,0 +1,1 @@
+error: `else` found outside `If` block (at offset 0x18)
