+++ tests/cli/print-no-panic-dangling-else.wat.stdout
@@ -1,6 +1,3 @@
 (module
   (type (;0;) (func))
   (func (;0;) (type 0)
-  else
-  )
-)
