+++ tests/cli/print-no-panic-double-end.wat
@@ -1,4 +1,4 @@
-;; RUN: print %
+;; FAIL: print %
 
 (module
     (func end)
