+++ tests/cli/print-no-panic-double-end.wat.stdout
@@ -2,5 +2,3 @@
   (type (;0;) (func))
   (func (;0;) (type 0)
     end
-  )
-)
