+++ tests/cli/print-with-too-many-ends.wat
@@ -1,4 +1,4 @@
-;; RUN: print %
+;; FAIL: print %
 
 (module
   (func end i32.const 0 drop end)
