+++ tests/cli/print-with-too-many-ends.wat.stderr
@@ -1,0 +1,1 @@
+error: operators remaining after end of function body or expression (at offset 0x18)
