+++ tests/cli/print-with-too-many-ends.wat.stdout
@@ -2,8 +2,3 @@
   (type (;0;) (func))
   (func (;0;) (type 0)
     end
-    i32.const 0
-    drop
-    end
-  )
-)
