+++ tests/snapshots/cli/missing-features/custom-page-sizes.wast.json
@@ -2,7 +2,7 @@
   "source_filename": "tests/cli/missing-features/custom-page-sizes.wast",
   "commands": [
     {
-      "type": "assert_malformed",
+      "type": "assert_invalid",
       "line": 4,
       "filename": "custom-page-sizes.0.wasm",
       "module_type": "binary",
